requirements.txt
flask==3.0.3
Werkzeug==3.0.3
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0

